Rename DestinationModel to DirectionModel.
