+++ nextbike/models/__init__.py
@@ -1,7 +1,5 @@
+from nextbike.models.DirectionModel import DestinationModel
 from nextbike.models.DurationModel import DurationModel
-
-from nextbike.models.DestinationModel import DestinationModel
-
 from nextbike.models.utils import (
     duration_preparation,
     classification_preparation,
